myapp/requirements.txt
Flask==2.3.3
Flask_SQLAlchemy==3.0.3
python-dotenv==1.0.0
google-auth==2.23.0
